+++ index.html
@@ -12,98 +12,6 @@
             <section class="news-bar-container"></section>
             <section class="news-press-display"></section>
         </div>
-        <!-- <header>
-            <div class="title-container">
-                <div class="title-icon">
-                    <img src="./asset/icons/newsstand.svg" alt="뉴스스탠드" />
-                </div>
-                <h1 class="display-bold24">뉴스스탠드</h1>
-            </div>
-            <div class="display-medium16 date"></div>
-        </header>
-        <section class="news-bar-container">
-            <div class="news-bar-item">
-                <div class="news-bar-rolling"></div>
-            </div>
-            <div class="news-bar-item">
-                <div class="news-bar-rolling"></div>
-            </div>
-        </section>
-        <section class="news-press-display">
-            <div class="news-press-tab-container">
-                <div class="news-press-type-menu">
-                    <div class="news-press-tab-menu-active selected-bold16">
-                        전체 언론사
-                    </div>
-                    <div class="available-medium16">내가 구독한 언론사</div>
-                </div>
-                <div class="news-press-view-menu">
-                    <button class="news-display-list-button">
-                        <svg
-                            width="24"
-                            height="24"
-                            viewBox="0 0 24 24"
-                            fill="currentColor"
-                            xmlns="http://www.w3.org/2000/svg"
-                        >
-                            <path
-                                d="M19 5V19H5V5H19ZM20.1 3H3.9C3.4 3 3 3.4 3 3.9V20.1C3 20.5 3.4 21 3.9 21H20.1C20.5 21 21 20.5 21 20.1V3.9C21 3.4 20.5 3 20.1 3ZM11 7H17V9H11V7ZM11 11H17V13H11V11ZM11 15H17V17H11V15ZM7 7H9V9H7V7ZM7 11H9V13H7V11ZM7 15H9V17H7V15Z"
-                                fill="inherit"
-                            />
-                        </svg>
-                    </button>
-                    <button
-                        class="news-display-grid-button news-display-active"
-                    >
-                        <svg
-                            width="24"
-                            height="24"
-                            viewBox="0 0 24 24"
-                            fill="currentColor"
-                            xmlns="http://www.w3.org/2000/svg"
-                        >
-                            <path
-                                d="M3 11V3H11V11H3ZM3 21V13H11V21H3ZM13 11V3H21V11H13ZM13 21V13H21V21H13ZM5 9H9V5H5V9ZM15 9H19V5H15V9ZM15 19H19V15H15V19ZM5 19H9V15H5V19Z"
-                                fill="inherit"
-                            />
-                        </svg>
-                    </button>
-                </div>
-            </div>
-            <div class="news-display-container">
-                <ul class="news-press-grid-view"></ul>
-                <div class="news-press-list-view hidden">
-                    <nav
-                        class="list-view-category-bar available-medium14"
-                    ></nav>
-                    <div class="list-view-main-container"></div>
-                </div>
-                <button class="grid-left-arrow-button">
-                    <img
-                        src="./asset/icons/left-button.png"
-                        alt="이전 페이지"
-                    />
-                </button>
-                <button class="grid-right-arrow-button">
-                    <img
-                        src="./asset/icons/right-button.png"
-                        alt="다음 페이지"
-                    />
-                </button>
-                <button class="list-left-arrow-button hidden">
-                    <img
-                        src="./asset/icons/left-button.png"
-                        alt="이전 페이지"
-                    />
-                </button>
-                <button class="list-right-arrow-button hidden">
-                    <img
-                        src="./asset/icons/right-button.png"
-                        alt="다음 페이지"
-                    />
-                </button>
-            </div>
-        </section> -->
         <script src="./src/app.js" type="module"></script>
     </body>
 </html>
