+++ src/app.js
@@ -1,66 +1,9 @@
-// import getDate from "./header/getDate.js";
-// import { handleClickTitleIcon } from "./header/handleTitleIcon.js";
-// import { setGridView } from "./news-display/gridView.js";
-// import { setNewsBarRolling, infiniteRolling } from "./news-bar/rolling.js";
-// import { handleClickViewIcon } from "./news-display/setViewMode.js";
-// import { setCategories } from "./news-display/setCategory.js";
-// import { setListView } from "./news-display/listView.js";
-// import { initNewsPressData } from "./news-display/initNewsPressData.js";
-// import {
-//     handleClickGridNextButton,
-//     handleClickGridPrevButton,
-//     handleClickListNextButton,
-//     handleClickListPrevButton,
-// } from "./news-display/handlePageButton.js";
-
-import Component from "./core/Component.js";
 import Header from "./header/Header.js";
 import NewsBar from "./news-bar/NewsBar.js";
 import NewsDisplay from "./news-display/NewsDisplay.js";
 
-// let newsPressData = [];
-// export let gridPage = 1;
-// export let pageIndex = 0;
-
-// const setGridPage = (num) => {
-//     gridPage = num;
-// };
-// const setListPageIndex = (num) => {
-//     pageIndex = num;
-// };
-
-// (async function () {
-//     newsPressData = await initNewsPressData();
-//     // header
-//     handleClickTitleIcon();
-//     getDate();
-
-//     // news bar rolling
-//     setNewsBarRolling();
-//     infiniteRolling();
-
-//     // news display
-//     handleClickViewIcon();
-
-//     setGridView();
-
-//     setCategories();
-//     // setListView();
-
-//     handleClickGridPrevButton();
-//     handleClickGridNextButton();
-//     handleClickListPrevButton();
-//     handleClickListNextButton();
-// })();
-
-// export { newsPressData, setGridPage, setListPageIndex };
-
-// export default function App() {
 const $app = document.querySelector(".root");
 
 new Header($app.querySelector(".news-stand-header"));
-
 new NewsBar($app.querySelector(".news-bar-container"));
-
 new NewsDisplay($app.querySelector(".news-press-display"));
-// }
