--- src/utils/insertHTML.js
@@ -1,3 +1,0 @@
-export default function insertHTML(target, template) {
-    return target.insertAdjacentHTML("beforeend", template);
-}
